+++ deltafi-ui/src/pages/DeltaFileUploadPage.vue
@@ -25,6 +25,19 @@
           <template #header>
             <Button label="Add Metadata Field" icon="pi pi-plus" class="mr-2" @click="addMetadataField" />
             <Button label="Clear" icon="pi pi-times" :disabled="metadataClearDisabled" class="mr-2" @click="clearMetadata" />
+            <div class="float-right btn-group">
+              <Button label="Export Metadata" icon="fas fa-download fa-fw" class="p-button-md p-button-secondary p-button-outlined mx-1" :disabled="metadataClearDisabled" @click="onExportMetadata" />
+              <FileUpload ref="metadataFileUploader" auto mode="basic" choose-label="Import Metadata" accept=".json,application/JSON" :file-limit="1" custom-upload class="p-button-md p-button-secondary p-button-outlined mx-1" @uploader="preUploadMetadataValidation" @click="setOverlayPanelPosition" />
+              <OverlayPanel ref="errorOverlayPanel" dismissable show-close-icon @hide="clearMetadataUploadErrors">
+                <Message severity="error" :sticky="true" class="mb-2 mt-0" :closable="false">
+                  <ul>
+                    <div v-for="(error, key) in _.uniq(errorsList)" :key="key">
+                      <li class="text-wrap text-break">{{ error }}</li>
+                    </div>
+                  </ul>
+                </Message>
+              </OverlayPanel>
+            </div>
           </template>
           <div class="row">
             <div class="col-5 p-fluid">
@@ -111,6 +124,8 @@ import Dropdown from "primevue/dropdown";
 import FileUpload from "primevue/fileupload";
 import InputText from "primevue/inputtext";
 import Menu from "primevue/menu";
+import Message from "primevue/message";
+import OverlayPanel from "primevue/overlaypanel";
 import Panel from "primevue/panel";
 import ProgressBar from "primevue/progressbar";
 import CollapsiblePanel from "@/components/CollapsiblePanel";
@@ -118,8 +133,10 @@ import MetadataViewer from "@/components/MetadataViewer.vue";
 import PageHeader from "@/components/PageHeader.vue";
 import DidLink from "@/components/DidLink.vue";
 import Timestamp from "@/components/Timestamp.vue";
+import useMetadataConfiguration from "@/composables/useMetadataConfiguration";
 import useFlows from "@/composables/useFlows";
 import useIngress from "@/composables/useIngress";
+import useNotifications from "@/composables/useNotifications";
 import { useStorage, StorageSerializers } from "@vueuse/core";
 import { ref, computed, onBeforeMount, watch } from "vue";
 import _ from "lodash";
@@ -132,6 +149,13 @@ const fileUploader = ref();
 const deltaFiles = ref([]);
 const { fetchActiveIngressFlows, activeIngressFlows } = useFlows();
 const { ingressFile } = useIngress();
+const notify = useNotifications();
+const { validateMetadataFile } = useMetadataConfiguration();
+const metadataFileUploader = ref(null);
+const errorsList = ref([]);
+const validUpload = ref({});
+const errorOverlayPanel = ref(null);
+const overlayPanelPosition = ref({});
 
 const deltaFilesMenuItems = ref([
   {
@@ -189,7 +213,7 @@ const clearMetadata = () => {
 };
 
 const onUpload = (event) => {
-    ingressFiles(event);
+  ingressFiles(event);
 };
 
 const ingressFiles = async (event) => {
@@ -212,7 +236,7 @@ watch(
     }
   },
   { deep: true }
-)
+);
 
 // Store for the sessions user selected Flow.
 const selectedFlowStorage = useStorage("selectedFlowStorage-session-storage", {}, sessionStorage, { serializer: StorageSerializers.object });
@@ -220,6 +244,14 @@ const selectedFlowStorage = useStorage("selectedFlowStorage-session-storage", {}
 const metadataStorage = useStorage("metadataStorage-session-storage", {}, sessionStorage, { serializer: StorageSerializers.object });
 // Store for the sessions user uploaded deltaFiles.
 const deltaFilesStorage = useStorage("deltafiles-upload-session-storage", {}, sessionStorage, { serializer: StorageSerializers.object });
+
+const storeMetaDataUploadSession = async () => {
+  // Save off inputed metadata into store.
+  metadataStorage.value = metadata.value;
+
+  // Save off selected flow into store.
+  selectedFlowStorage.value = selectedFlow.value;
+};
 
 const storeDeltaFileUploadSession = async (results) => {
   // If there is no data in the deltaFiles storage then just save it off. If data is in there we want to persist it so concat the older data with
@@ -259,8 +291,9 @@ const getSelectedFlowSession = () => {
 
 const replayMetadata = (value) => {
   metadata.value = JSON.parse(JSON.stringify(value.uploadedMetadata));
-  let flowSelected = `{"name" : "${value.flow}"}`;
-  selectedFlow.value = JSON.parse(flowSelected);
+  let flowSelected = {};
+  flowSelected["name"] = value.flow;
+  selectedFlow.value = flowSelected;
 };
 
 const clearDeltaFilesSession = () => {
@@ -279,6 +312,90 @@ const formatMetadataforViewer = (filename, uploadedMetadata) => {
   let metaDataObject = {};
   metaDataObject[filename] = uploadedMetadata;
   return JSON.parse(JSON.stringify(metaDataObject));
+};
+
+const onExportMetadata = () => {
+  let formattedMetadata = {};
+  if (_.isEmpty(selectedFlow.value) && _.isEmpty(metadataRecord.value)) {
+    return;
+  }
+  if (!_.isEmpty(selectedFlow.value)) {
+    formattedMetadata["flow"] = selectedFlow.value.name;
+  }
+  formattedMetadata["metadata"] = metadataRecord.value;
+
+  exportMetadataFile(formattedMetadata);
+};
+
+const exportMetadataFile = (formattedMetadata) => {
+  let link = document.createElement("a");
+  let downloadFileName = "metadata_export_" + new Date(Date.now()).toLocaleDateString();
+  link.download = downloadFileName.toLowerCase();
+  let blob = new Blob([JSON.stringify(formattedMetadata, null, 2)], {
+    type: "application/json",
+  });
+  link.href = URL.createObjectURL(blob);
+  link.click();
+  URL.revokeObjectURL(link.href);
+  link.remove();
+};
+
+const preUploadMetadataValidation = async (request) => {
+  for (let file of request.files) {
+    let reader = new FileReader();
+
+    reader.readAsText(file);
+
+    reader.onload = function () {
+      let uploadNotValid = validateMetadataFile(reader.result);
+      errorsList.value = [];
+      if (uploadNotValid) {
+        for (let errorMessages of uploadNotValid) {
+          errorsList.value.push(errorMessages.message);
+        }
+        deleteMetadataFile();
+        errorOverlayPanel.value.toggle(overlayPanelPosition.value, overlayPanelPosition.value.target);
+      } else {
+        validUpload.value = reader.result;
+        uploadMetadataFile(file);
+      }
+    };
+  }
+};
+
+const uploadMetadataFile = async (file) => {
+  let parseMetadataUpload = JSON.parse(validUpload.value);
+  let flowSelected = {};
+  if (!_.isEmpty(_.get(parseMetadataUpload, "flow"))) {
+    flowSelected["name"] = _.get(parseMetadataUpload, "flow");
+    selectedFlow.value = flowSelected;
+  } else {
+    selectedFlow.value = null;
+  }
+  let reformatMetadata = [];
+  for (const [key, value] of Object.entries(parseMetadataUpload.metadata)) {
+    let formatMetadata = {};
+    formatMetadata["key"] = key;
+    formatMetadata["value"] = value;
+    reformatMetadata.push(formatMetadata);
+  }
+  metadata.value = reformatMetadata;
+  notify.success(`Uploaded ${file.name}`, `Successfully imported metadata`, 4000);
+  storeMetaDataUploadSession();
+  deleteMetadataFile();
+};
+
+const deleteMetadataFile = () => {
+  metadataFileUploader.value.files = [];
+  metadataFileUploader.value.uploadedFileCount = 0;
+};
+
+const clearMetadataUploadErrors = () => {
+  errorsList.value = [];
+};
+
+const setOverlayPanelPosition = (event) => {
+  overlayPanelPosition.value = event;
 };
 </script>